Fix mic not using correct bus.
In the demo, it assumes the mic is on the Capture bus, but after importing
a project may not (and should need) to have its audio buses setup for the
demo, so Godot resets it to Master if Capture doesn't exist.

During _init, the demo loads the demo bus layout, so the mic is now changed
to use the "Capture" bus in _ready.